+++ app/assets/javascripts/service_uploads/ServiceTypeDropdown.js
@@ -1,4 +1,4 @@
-
+import React from 'react';
 class ServiceTypeDropdown extends React.Component {
 
   componentDidMount() {
